UIST/2/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
reportlab==4.0.7
openpyxl==3.1.2
Werkzeug==2.3.7
Pillow==10.1.0